pyproject.toml
[project]
name = "gyaan-doc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "langchain-core>=0.3.61",
    "langchain-google-genai>=2.1.4",
    "langgraph>=0.4.5",
    "python-dotenv>=1.1.0",
]
